@@ -3,13 +3,16 @@ import { ShoppingCartService } from "app/restaurant-detail/shopping-cart/shoppin
 import { CartItem } from "app/restaurant-detail/shopping-cart/cart-item.model";
 import { Order } from "./order.model";
 import { Observable } from "rxjs/Observable";
-import { Http, RequestOptions, Headers } from "@angular/http";
+import { HttpClient } from "@angular/common/http";
 import { MEAT_API } from "app/app.api";
-import 'rxjs/add/operator/map'
+import "rxjs/add/operator/map";
 
 @Injectable()
 export class OrderService {
-  constructor(private cartService: ShoppingCartService, private http: Http) {}
+  constructor(
+    private cartService: ShoppingCartService,
+    private http: HttpClient
+  ) {}
 
   itemsValue(): number {
     return this.cartService.total();
@@ -31,21 +34,14 @@ export class OrderService {
     this.cartService.removeItem(item);
   }
 
-  clear(){
-    this.cartService.clear()
+  clear() {
+    this.cartService.clear();
   }
 
   /* Como é uma chamada Http então utiliza-se Observable */
   checkOrder(order: Order): Observable<string> {
-    const headers = new Headers();
-    headers.append("Content-Type", "application/json");
     return this.http
-      .post(
-        `${MEAT_API}/orders`,
-        JSON.stringify(order),
-        new RequestOptions({headers: headers})
-      )
-      .map(response => response.json())
+      .post<Order>(`${MEAT_API}/orders`, order)
       .map(order => order.id);
   }
 }
