@@ -1,7 +1,7 @@
 /* Esta é a classe que acede ao backend */
 
 import { Injectable } from "@angular/core";
-import { Http } from "@angular/http";
+import { HttpClient, HttpParams } from "@angular/common/http";
 
 import { Observable } from "rxjs/Observable";
 import "rxjs/add/operator/map";
@@ -15,38 +15,30 @@ import { MenuItem } from "app/restaurant-detail/menu-item/menu-item.model";
 
 @Injectable()
 export class RestaurantsService {
-  constructor(private http: Http) {}
+  constructor(private http: HttpClient) {}
 
   /* Permitir a pesquisa (parametro search) */
   restaurants(search?: string): Observable<Restaurant[]> {
-    return this.http
-      .get(`${MEAT_API}/restaurants`, {
-        params: { q: search }
-      }) /* o q permite que a (query) pesquisa seja feita em todos os elemtos | json serve permite isso */
-      .map(response => response.json())
-      .catch(
-        ErrorHandler.handleError
-      ); /* utilizado para exibir erro no console | vai utilizar o metodo handleError da função ErrorHandler */
+    let params: HttpParams = undefined;
+    if (search) {
+      params = new HttpParams().append("q", search);
+    }
+
+    return this.http.get<Restaurant[]>(`${MEAT_API}/restaurants`, {
+      params: params
+    });
+    /* o q permite que a (query) pesquisa seja feita em todos os elemtos | json serve permite isso */
   }
 
   restaurantById(id: string): Observable<Restaurant> {
-    return this.http
-      .get(`${MEAT_API}/restaurants/${id}`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError);
+    return this.http.get<Restaurant>(`${MEAT_API}/restaurants/${id}`);
   }
 
   reviewsOfRestaurant(id: string): Observable<any> {
-    return this.http
-      .get(`${MEAT_API}/restaurants/${id}/reviews`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError);
+    return this.http.get(`${MEAT_API}/restaurants/${id}/reviews`);
   }
 
   menuOfRestaurant(id: string): Observable<MenuItem[]> {
-    return this.http
-      .get(`${MEAT_API}/restaurants/${id}/menu`)
-      .map(response => response.json())
-      .catch(ErrorHandler.handleError);
+    return this.http.get<MenuItem[]>(`${MEAT_API}/restaurants/${id}/menu`);
   }
 }
